requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
pyarrow==15.0.2
pyyaml==6.0.1
joblib==1.4.2
torch==2.2.2
mlflow==2.13.0
dask[dataframe]==2024.5.2
pytest==8.2.0
optuna==3.6.0